+++ app/models/sublease/lodger.rb
@@ -1,7 +1,8 @@
+# frozen_string_literal: true
+
 module Sublease
   class Lodger < ActiveRecord::Base
     self.abstract_class = true
-    attr_accessor :allow_sublease_change
 
     class << self
 
@@ -16,11 +17,17 @@ module Sublease
       def belongs_to_sublease(tenant, options = {})
         belongs_to tenant.to_sym, options.delete(:dependent)
 
-        before_validation -> {add_sublease_error("#{tenant}_id".to_sym)}, on: :update, if: "#{tenant}_id_changed?".to_sym
+        before_validation -> {valid_sublease_change(tenant)}, on: :update
 
         validates "#{tenant}_id".to_sym, presence: true
+        validate -> {valid_sublease_tenant(tenant)}
 
       end
+    end
+
+    # Allow the sublease to be changed if true. Returns true or false.
+    def allow_sublease_change
+      @allow_sublease_change ||= false
     end
 
     # Allows the sublease to be changed. Set using true or false.
@@ -30,7 +37,7 @@ module Sublease
 
     # Checks if the sublease can be changed. Returns true or false.
     def allow_sublease_change?
-      allow_sublease_change.nil? ? false : allow_sublease_change
+      allow_sublease_change
     end
 
     # Overrides standard object delete to call destroy instead ensuring all callbacks are fired.
@@ -40,14 +47,23 @@ module Sublease
 
     private
 
-    def add_sublease_error(tenant)
-      unless allow_sublease_change?
-        msg = I18n.t('sublease.errors.changed')
-        if Rails::VERSION::MAJOR > 4
-          errors.add tenant, :changed, message: msg
-        else
-          errors.add tenant, message: msg
-        end
+    def add_sublease_error(tenant, type, msg = nil)
+      options = {}
+      unless msg.nil?
+        options = {message: msg}
+      end
+      errors.add tenant, type, options
+    end
+
+    def valid_sublease_change(tenant)
+      if (send("#{tenant}_id_changed?".to_sym) && (!allow_sublease_change?))
+        add_sublease_error("#{tenant}_id".to_sym, :changed, I18n.t('sublease.errors.changed'))
+      end
+    end
+
+    def valid_sublease_tenant(tenant)
+      unless tenant.to_s.capitalize.constantize.where(id: self.send("#{tenant}_id".to_sym)).exists?
+        add_sublease_error("#{tenant}_id".to_sym, :invalid)
       end
     end