+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.1.0.rc3
+-----
+* Refactor error messages and validations in Sublease::Lodger.
+* Changed before_filter to before_action to be compatible with Rails 5.1.
+* Added frozen_string_literal: true to all files that need it.
+* Added tests.
+
 0.1.0.rc2
 -----
 * Fixed issues with Sublease::TenantSwitcher not loading.